It's working!
Can talk to Peakflow and send commands via the cliRun. Not too advanced,
but hey - it's working! :D

Built using the ZSI SOAP framework which I'm not too happy with but it
is working with digesting a WSDL and doing HTTP digest authentication so
I'll settle with this for now.

Naturally, ZSI was broken too, so had to monkey patch my way through the
broken function.